+++ cbsyst/cbsyst.py
@@ -52,9 +52,11 @@ def Csys(
     pH, DIC, CO2, HCO3, CO3, TA : array-like
         Carbon system parameters. Two of these must be provided.
     BT : array-like
-        Total B at Salinity = 35, used in Alkalinity calculations.
-    Ca, Mg : arra-like
-        The [Ca] and [Mg] of the seawater, in mol / kg.
+        Total B at the input salinity used in Alkalinity calculations. 
+        If missing, this is calculated as 0.000416 * Sal/35 
+        (Uppstrom et al. 1974).
+    Ca, Mg : array-like
+        The [Ca] and [Mg] of seawater, in mol / kg.
         Used in calculating MyAMI constants.
     T_in, S_in : array-like
         Temperature in Celcius and Salinity in PSU that the
@@ -123,7 +125,7 @@ def Csys(
         ps.BT = calc_TB(ps.S_in)
 
     # Remove negative values 
-    for p in ["DIC", "TA", "CO2", "HCO3", "CO3", "BT", "fCO2", "pCO2", "TP", "TSi"]:
+    for p in ["DIC", "CO2", "HCO3", "CO3", "BT", "fCO2", "pCO2", "TP", "TSi"]:
         if ps[p] is not None:
             if isinstance(ps[p], (np.ndarray, pd.core.series.Series)):
                 ps[p][ps[p] < 0] = np.nan
@@ -162,12 +164,20 @@ def Csys(
         TF=ps.TF,
         Ks=ps.Ks,
     )
+    
+    # calc Omega
+    if Ca is None:
+        oCa = 0.0102821 * ps.S_in / 35.
+    else:
+        oCa = Ca * ps.S_in / 35.
+    ps['OmegaA'] = ps['CO3'] * oCa / ps.Ks.KspA
+    ps['OmegaC'] = ps['CO3'] * oCa / ps.Ks.KspC
 
     # clean up output
     outputs = [
         "BT", "CO2", "CO3", "Ca", "DIC", "H", "HCO3", 
         "Mg", "S_in", "T_in", "TA", "CAlk", "PAlk", 
-        "SiAlk", "OH"]
+        "SiAlk", "OH", 'OmegaA', 'OmegaC', 'revelle_factor']
     for k in outputs:
         if not isinstance(ps[k], np.ndarray):
             # convert all outputs to (min) 1D numpy arrays.
@@ -183,9 +193,20 @@ def Csys(
             ps.T_out = ps.T_in
         if ps.S_out is None:
             ps.S_out = ps.S_in
+        else:
+            # if salinity is modified, update salinity-dependent parameters
+            BT = ps.BT * ps.S_out / ps.S_in
+            TS = ps.TS * ps.S_out / ps.S_in
+            TF = ps.TF * ps.S_out / ps.S_in
+            if Ca is not None:
+                Ca = ps.Ca * ps.S_out / ps.S_in
+            if Mg is not None:
+                Mg = ps.Mg * ps.S_out / ps.S_in
         if ps.P_out is None:
             ps.P_out = ps.P_in
         # assumes conserved alkalinity and DIC
+        
+        # This needs to be different depending on whether T or S is changing.
         out_cond = Csys(
             TA=ps.TA,
             DIC=ps.DIC,
@@ -193,6 +214,11 @@ def Csys(
             S_in=ps.S_out,
             P_in=ps.P_out,
             unit=ps.unit,
+            Ca=Ca,
+            Mg=Mg,
+            BT=BT,
+            TF=TF,
+            TS=TS,
         )
 
         # rename parameters in output conditions
@@ -200,7 +226,7 @@ def Csys(
             "BAlk", "BT", "CAlk", "CO2", "CO3", "DIC", "H", "HCO3", 
             "HF", "HSO4", "Hfree", "Ks", "OH", "PAlk", "SiAlk", "TA", "TF",
             "TP", "TS", "TSi", "fCO2", "pCO2", "pHfree", "pHsws", 
-            "pHtot", "pHNBS", "revelle_factor",
+            "pHtot", "pHNBS", 'OmegaA', 'OmegaC', "revelle_factor",
         ]
 
         ps.update({k + "_in": ps[k] for k in outputs})
@@ -700,7 +726,7 @@ def CBsys(
         ps.TF = calc_TF(ps.S_in)
 
     # Remove negative values 
-    for p in ["DIC", "TA", "CO2", "HCO3", "CO3", "BT", "BO3", "BO4", "fCO2", "pCO2", "TP", "TSi"]:
+    for p in ["DIC", "CO2", "HCO3", "CO3", "BT", "BO3", "BO4", "fCO2", "pCO2", "TP", "TSi"]:
         if ps[p] is not None:
             if isinstance(ps[p], (np.ndarray, pd.core.series.Series)):
                 ps[p][ps[p] < 0] = np.nan
@@ -805,6 +831,14 @@ def CBsys(
         Ks=ps.Ks,
     )
 
+    # calc Omega
+    if Ca is None:
+        oCa = 0.0102821 * ps.S_in / 35.
+    else:
+        oCa = Ca * ps.S_in / 35.
+    ps['OmegaA'] = ps['CO3'] * oCa / ps.Ks.KspA
+    ps['OmegaC'] = ps['CO3'] * oCa / ps.Ks.KspC
+    
     # clean up output
     outputs = [
         "BAlk",
@@ -833,6 +867,9 @@ def CBsys(
         "pHsws",
         "pHtot",
         "pHNBS",
+        'OmegaA',
+        'OmegaC',
+        'revelle_factor',
         "BO3",
         "BO4",
         "ABO3",
@@ -861,18 +898,31 @@ def CBsys(
             ps.T_out = ps.T_in
         if ps.S_out is None:
             ps.S_out = ps.S_in
+        else:
+            # if salinity is modified, update salinity-dependent parameters
+            BT = ps.BT * ps.S_out / ps.S_in
+            TS = ps.TS * ps.S_out / ps.S_in
+            TF = ps.TF * ps.S_out / ps.S_in
+            if Ca is not None:
+                Ca = ps.Ca * ps.S_out / ps.S_in
+            if Mg is not None:
+                Mg = ps.Mg * ps.S_out / ps.S_in
         if ps.P_out is None:
             ps.P_out = ps.P_in
         # assumes conserved alkalinity, DIC and BT
         out_cond = CBsys(
             TA=ps.TA,
             DIC=ps.DIC,
-            BT=ps.BT,
             dBT=ps.dBT,
             T_in=ps.T_out,
             S_in=ps.S_out,
             P_in=ps.P_out,
             unit=ps.unit,
+            Ca=Ca,
+            Mg=Mg,
+            BT=BT,
+            TF=TF,
+            TS=TS,
         )
         # rename parameters in output conditions
         ps.update({k + "_in": ps[k] for k in outputs})